+++ src/bigstring.mli
@@ -357,6 +357,9 @@ val set_head_padded_fixed_string : padding:char -> t -> pos:int -> len:int -> st
 
 
 
+(*_ See the Jane Street Style Guide for an explanation of [Private] submodules:
+
+  https://opensource.janestreet.com/standards/#private-submodules *)
 module Private : sig
   val sign_extend_16 : int -> int
 end