+++ src/thread_safe_queue.mli
@@ -46,6 +46,9 @@ val dequeue_exn : 'a t -> 'a
     by the garbage collector. *)
 val clear_internal_pool : _ t -> unit
 
+(*_ See the Jane Street Style Guide for an explanation of [Private] submodules:
+
+  https://opensource.janestreet.com/standards/#private-submodules *)
 module Private : sig
   module Uopt : sig
     type 'a t [@@deriving sexp_of]